testing scaffoling for client (jest, enzyme) and server (jest, supertest)
