+++ .travis.yml
@@ -1,0 +1,6 @@
+language: node_js
+node_js:
+  - 10
+
+script:
+  - yarn test
