+++ client/components/App/index.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const App = () => {
+    return <div>Hello React!</div>;
+};
+
+export default App;
