+++ client/index.js
@@ -1,8 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-
-const App = () => {
-  return <div>Hello React!</div>;
-};
+import App from '@client/components/App';
 
 ReactDOM.render(<App />, document.getElementById("root"));
