+++ client/tests/App.test.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Enzyme, { shallow } from 'enzyme';
+import EnzymeAdapter from 'enzyme-adapter-react-16';
+
+import App from '../components/App';
+
+Enzyme.configure({ adapter: new EnzymeAdapter() });
+
+test('renders without crashing', () => {
+    const wrapper = shallow(<App />);
+    expect(wrapper).toBeTruthy();
+});
+
