+++ client/webpack.config.js
@@ -4,6 +4,8 @@ const CopyWebpackPlugin = require('copy-webpack-plugin');
 
 module.exports = {
     entry: ['babel-polyfill', './index.js'],
+    mode: 'development',
+    devtool: 'inline-source-map',
     module: {
         rules: [
             {
