+++ docker-compose.test.yml
@@ -1,0 +1,7 @@
+version: '3'
+
+services:
+    api:
+        command: "yarn test --watchAll"
+    app:
+        command: "yarn test --watchAll"
