+++ package.json
@@ -5,7 +5,9 @@
   "main": "server/index.js",
   "private": true,
   "scripts": {
-    "dev": "docker-compose up"
+    "dev": "docker-compose up",
+    "test": "yarn workspaces run test",
+    "testAll": "docker-compose -f docker-compose.yml -f docker-compose.test.yml up"
   },
   "repository": {
     "type": "git",
